requirements.txt
pandas~=2.2.3
numpy~=2.1.3
rich~=13.9.4

setuptools~=68.2.0
torch~=2.5.1
opencv-python~=4.10.0.84
pillow~=11.0.0
matplotlib~=3.10.0